refactor: live preivew custom server tests to seperate suite
